backend/requirements.txt
# Ghana Healthcare Intelligence Platform - Backend Dependencies

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Databricks
databricks-sql-connector==3.0.2
openai==1.10.0

# Data processing
numpy==1.26.3

# Environment variables
python-dotenv==1.0.0

# Optional: Production deployment
gunicorn==21.2.0
